1. Script 의 맴버를 ScriptUI 에 노출시켜서 런타임 도중에 변경 가능하도록 함

+++ Project/Client/ScriptUI.h
@@ -22,6 +22,7 @@ public:
 
 private:
 	class qScript*			m_Script;
-	uintptr_t				m_UIHeight;
+
+	UINT					m_UIHeight;
 };
